-addition of VS2019 solution and project files, and removal of VS2013 solutions and projects
-replace std::auto_ptr with std::unique_ptr
-return project, replace boost::shared_ptr with std::shared_ptr
-replace resip::SharedPtr with std::shared_ptr
-makefile now requires C++11 compiler support
-replace boost::bind with std::bind or lambda expressions

+++ rutil/hep/HepAgent.hxx
@@ -162,7 +162,7 @@ class HepAgent
          hg->time_sec.chunk.vendor_id = htons(0x0000);
          hg->time_sec.chunk.type_id   = htons(0x0009);
          hg->time_sec.chunk.length = htons(sizeof(hg->time_sec));
-         hg->time_sec.data = htonl(now / 1000000LL);
+         hg->time_sec.data = htonl((u_long)(now / 1000000LL));
 
          /* TIMESTAMP USEC */
          hg->time_usec.chunk.vendor_id = htons(0x0000);
